refactor: separate target outside trampoline.rs
